refactor: split combat.ts and convert effects to registry-based dispatch
Phase 1: Split validActions/combat.ts (1,029 lines) into domain files
- combatHelpers.ts - Shared utilities (resistances, elemental calc)
- combatAttack.ts - Attack pool and RANGED_SIEGE/ATTACK phases
- combatBlock.ts - Block pool and BLOCK phase
- combatDamage.ts - Damage assignment and ASSIGN_DAMAGE phase
- combat.ts now just contains getCombatOptions dispatcher

Phase 2: Convert effects/index.ts to Map-based dispatch
- effectRegistry.ts - Map-based dispatch with registerEffect()
- effectRegistrations.ts - Registers all 21 effect handlers
- Removed 31-case switch statement from resolveEffect()

Benefits:
- Reduced merge conflict risk in frequently-modified files
- Enables parallel development on different domains
- New effects/combat phases can be added without modifying central files
